Extricate the swaps
