+++ cirq/ops/swap_gates_test.py
@@ -1,0 +1,123 @@
+# Copyright 2019 The Cirq Developers
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+import pytest
+import sympy
+
+import cirq
+
+
+@pytest.mark.parametrize('eigen_gate_type', [
+    cirq.ISwapPowGate,
+    cirq.SwapPowGate,
+])
+def test_phase_sensitive_eigen_gates_consistent_protocols(eigen_gate_type):
+    cirq.testing.assert_eigengate_implements_consistent_protocols(
+        eigen_gate_type, ignoring_global_phase=True)
+
+
+def test_text_diagrams():
+    a = cirq.NamedQubit('a')
+    b = cirq.NamedQubit('b')
+    circuit = cirq.Circuit(cirq.SWAP(a, b), cirq.ISWAP(a, b)**-1)
+
+    cirq.testing.assert_has_diagram(
+        circuit, """
+a: ───×───iSwap──────
+      │   │
+b: ───×───iSwap^-1───
+""")
+
+    cirq.testing.assert_has_diagram(circuit,
+                                    """
+a: ---swap---iSwap------
+      |      |
+b: ---swap---iSwap^-1---
+""",
+                                    use_unicode_characters=False)
+
+
+def test_swap_unitary():
+    # yapf: disable
+    np.testing.assert_almost_equal(
+        cirq.unitary(cirq.SWAP**0.5),
+        np.array([
+            [1, 0, 0, 0],
+            [0, 0.5 + 0.5j, 0.5 - 0.5j, 0],
+            [0, 0.5 - 0.5j, 0.5 + 0.5j, 0],
+            [0, 0, 0, 1]
+        ]))
+    # yapf: enable
+
+
+def test_iswap_unitary():
+    # yapf: disable
+    cirq.testing.assert_allclose_up_to_global_phase(
+        cirq.unitary(cirq.ISWAP),
+        # Reference for the iswap gate's matrix using +i instead of -i:
+        # https://quantumcomputing.stackexchange.com/questions/2594/
+        np.array([[1, 0, 0, 0],
+                   [0, 0, 1j, 0],
+                   [0, 1j, 0, 0],
+                   [0, 0, 0, 1]]),
+        atol=1e-8)
+    # yapf: enable
+
+
+def test_repr():
+    assert repr(cirq.SWAP) == 'cirq.SWAP'
+    assert repr(cirq.SWAP**0.5) == '(cirq.SWAP**0.5)'
+
+    assert repr(cirq.ISWAP) == 'cirq.ISWAP'
+    assert repr(cirq.ISWAP**0.5) == '(cirq.ISWAP**0.5)'
+
+
+def test_str():
+    assert str(cirq.SWAP) == 'SWAP'
+    assert str(cirq.SWAP**0.5) == 'SWAP**0.5'
+
+    assert str(cirq.ISWAP) == 'ISWAP'
+    assert str(cirq.ISWAP**0.5) == 'ISWAP**0.5'
+
+
+def test_iswap_trace_distance():
+    for exp in np.linspace(0, 4, 20):
+        cirq.testing.assert_has_consistent_trace_distance_bound(cirq.ISWAP**exp)
+
+
+def test_iswap_decompose_diagram():
+    a = cirq.NamedQubit('a')
+    b = cirq.NamedQubit('b')
+
+    decomposed = cirq.Circuit(cirq.decompose_once(cirq.ISWAP(a, b)**0.5))
+    cirq.testing.assert_has_diagram(
+        decomposed, """
+a: ───@───H───X───T───X───T^-1───H───@───
+      │       │       │              │
+b: ───X───────@───────@──────────────X───
+""")
+
+
+def test_trace_distance():
+    foo = sympy.Symbol('foo')
+    sswap = cirq.SWAP**foo
+    siswap = cirq.ISWAP**foo
+    # These values should have 1.0 or 0.0 directly returned
+    assert cirq.trace_distance_bound(sswap) == 1.0
+    assert cirq.trace_distance_bound(siswap) == 1.0
+    # These values are calculated, so we use approx_eq
+    assert cirq.approx_eq(cirq.trace_distance_bound(cirq.SWAP**0.3),
+                          np.sin(0.3 * np.pi / 2))
+    assert cirq.approx_eq(cirq.trace_distance_bound(cirq.ISWAP**0), 0.0)
